Refactor torch_npu
