--- torch_npu/npu/config.yaml
@@ -1,1 +1,0 @@
-flamegraph_url: 'https://raw.githubusercontent.com/brendangregg/FlameGraph/master/flamegraph.pl'
